% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_mutations.R
\name{apply_mutations}
\alias{apply_mutations}
\title{Apply random mutations to a community}
\usage{
apply_mutations(traits_comm, mutation_sd = default_mutation_sd())
}
\arguments{
\item{traits_comm}{trait values of individuals in the community}

\item{mutation_sd}{numeric \verb{>= 0}, the standard deviation of the normal
distribution from which mutations are drawn.}
}
\description{
The trait value of each individual in the input community is modified by a mutation sampled in a normal distribution
of mean \code{0} and standard deviation \code{mutation_sd}.
}
\author{
Theo Pannetier
}
