% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_speciation.R
\name{apply_speciation}
\alias{apply_speciation}
\title{Resolve speciation in a comrad community}
\usage{
apply_speciation(comm, trait_dist_sp = default_trait_dist_sp())
}
\arguments{
\item{comm}{a tibble with one row per individual in the community and three
columns:
\itemize{
\item \code{z} contains (numeric) traits values
\item \code{species} contains species names (characters)
\item \code{ancestral_species} contains acnestral species names (characters)
}}

\item{trait_dist_sp}{numeric, the minimal trait distance between two
clusters of individuals triggering speciation.}
}
\description{
For each species, check if there is any gap \verb{>= trait_dist_sp} in trait
values, and split the relevant species in two when one is found. The less
numerous half becomes the new species.
}
\note{
\code{apply_speciation()} can currently only split one species into two. If
multiple gaps emerge in a species at a single time step, only the first one
will be treated. As long as branching does not happen at every generation,
this is unlikely to be an issue, because an unresolved gap will be caught on
the next time step. In phylogenetic terms, this results in polytomies being
resolved as soft polytomies.
Besides, simultaneous branching events in different species are handled
perfectly fine.
}
\author{
Théo Pannetier
}
