% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_phylo_tbl.R
\name{assemble_phylo_tbl}
\alias{assemble_phylo_tbl}
\title{Extracts phylogenetic information from a \code{comrad} simulation}
\usage{
assemble_phylo_tbl(comrad_tbl)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}
}
\value{
a tibble with four columns and a row per species in the simulation
\itemize{
\item \code{species_name}, name of the species
\item \code{ancestor_name}, name of the ancestral species from which it branched. \code{NA}
for the initial species.
\item \code{time_birth}, the generation at which the species appeared.
\item \code{time_death}, the generation at which the last individual died. Last
generation if the species was still extant at the end of the simulation.
}
}
\description{
Reads the output of a \code{comrad} simulation and creates a table containing
the necessary information about every species to make a phylogeny.
}
\author{
Théo Pannetier
}
