% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carr_cap.R
\name{carr_cap}
\alias{carr_cap}
\title{Get the carrying capacity for a given trait value}
\usage{
carr_cap(trait, trait_opt = 0, carr_cap_opt = 1000,
  carr_cap_var = 0.2)
}
\arguments{
\item{trait}{numeric. Trait value \eqn{z}.}

\item{trait_opt}{numeric. Optimal value for the trait \eqn{z_{opt}}, at which
\eqn{K = K_{0}}}

\item{carr_cap_opt}{numeric. Maximum carrying capacity at \eqn{z = z_{opt}}.}

\item{carr_cap_var}{numeric. Variance of the carrying capacity
\eqn{\sigma^{2}_{K}}}
}
\description{
Computes the carrying capacity experienced by an individual with trait value
\code{trait}.
}
\author{
Theo Pannetier
}
