% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dd_model_is_avail.R
\name{check_dd_model_is_avail}
\alias{check_dd_model_is_avail}
\title{Check that the requested DD model is available in \code{DDD}}
\usage{
check_dd_model_is_avail(dd_model)
}
\arguments{
\item{dd_model}{a \code{dd_model} list, the output of e.g. \code{\link[=dd_model_lc]{dd_model_lc()}},
\code{\link[=dd_model_xc]{dd_model_xc()}}, etc.}
}
\description{
Calls \code{dd_sim()} from \code{DDD} to run a quick simulation and check
that the specified DD model is available on the version of DDD installed
on this computer.
}
