% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_coeff.R
\name{comp_coeff}
\alias{comp_coeff}
\title{Get the competition coefficient between two individuals}
\usage{
comp_coeff(trait_ind, trait_comp, sigma_comp = default_sigma_comp())
}
\arguments{
\item{trait_ind}{numeric. Trait value \eqn{z} of a focal individual.}

\item{trait_comp}{numeric, the trait value \eqn{z_j} of a competitor.}

\item{sigma_comp}{numeric > 0. Controls the competition intensity, i.e. the
strength of negative feedbacks the focal indiviudal experiences for a given
trait distance with the competitor.}
}
\description{
Returns the competition coefficient \eqn{\alpha(zi, z_j)} between a focal
individual \eqn{i}, and a competitor \eqn{j} with trait values \eqn{z_i},
\eqn{z_j}.
}
\details{
The equation is reported from equation (3) in Pontarp et al. (2012).
\deqn{\alpha(z_i, z_j) = exp(-(z_i - z_j) ^ 2 / (2 * \sigma^2_\alpha))}
}
\author{
Theo Pannetier
}
