% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sim_to_phylo.R
\name{convert_sim_to_phylo}
\alias{convert_sim_to_phylo}
\title{Convert \code{comrad} simulation results into a phylogeny}
\usage{
convert_sim_to_phylo(comrad_tbl, with_extinct = TRUE)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}

\item{with_extinct}{logical, should extinct taxa be included? \code{TRUE} returns
the full tree, \code{FALSE} returns the reconstructed tree.}
}
\description{
Reads a table of results and returns a \code{phylo} class object.
}
\author{
Théo Pannetier
}
