% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_newick.R
\name{convert_to_newick}
\alias{convert_to_newick}
\title{Convert a species history table to a Newick string}
\usage{
convert_to_newick(phylo_tbl)
}
\arguments{
\item{phylo_tbl}{a table with phylogenetic information for each species, the
output of \code{\link[=assemble_phylo_tbl]{assemble_phylo_tbl()}}}
}
\description{
Reads a tibble recording species birth, death and phylogenetic relations,
figures out the branch lenghts and return a phylogenetic tree in Newick
format. Based on \code{\link[DDD:L2phylo]{DDD::L2phylo()}}.
}
\author{
Théo Pannetier
}
