% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_next_gen_traits.R
\name{create_next_gen_traits}
\alias{create_next_gen_traits}
\title{Create offspring trait values from parent values}
\usage{
create_next_gen_traits(
  traits_pop,
  nb_offspring_pop,
  mutation_sd = default_mutation_sd()
)
}
\arguments{
\item{traits_pop}{numeric vector of variable length, the trait values of
every individual in the population.}

\item{nb_offspring_pop}{numeric vector \code{>= 0}, the nb of offspring that
should be created for each parent.}

\item{mutation_sd}{numeric \eqn{>= 0}, the standard deviation of the normal
distrbution in which the mutations are drawn.}
}
\description{
Each parent first sires a number of offspring equal to its
\code{nb_offspring_pop} value, then each offspring receives a mutation
sampled in a normal distribution of mean \code{0} and variance
\code{mutation_sd}.
}
\author{
Theo Pannetier
}
