% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_cc.R
\name{dd_model_cc}
\alias{dd_model_cc}
\title{Diversity-dependent model with constant-rate (c) speciation and constant-rate
(c) extinction}
\usage{
dd_model_cc()
}
\description{
A list specifying a DD model with constant-rate
speciation and constant-rate extinction; to be fed as argument
\code{dd_model} to \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
}
\details{
\deqn{\lambda(N) = \lambda_{0} \\ \mu(N) = \mu_{0}}
}
\author{
Theo Pannetier
}
