% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_cl.R
\name{dd_model_cl}
\alias{dd_model_cl}
\title{Diversity-dependent model with constant-rate (c) speciation and
linear (l) dependence on extinction}
\usage{
dd_model_cl()
}
\description{
A list specifying a DD model with linear diversity-dependence on the
extinction rate; to be passed as argument
\code{dd_model} to \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
}
\details{
\deqn{\lambda(N) = \lambda_{0}}
\deqn{\mu(N) = \mu_{0} + (\lambda_{0} - \mu_{0}) \frac{N}{K}}
}
\author{
Theo Pannetier
}
