% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_cp.R
\name{dd_model_cp}
\alias{dd_model_cp}
\title{Diversity-dependent model with constant-rate (c) speciation and
power (p) dependence on extinction}
\usage{
dd_model_cp()
}
\description{
A list specifying a DD model with constant-rate speciation
and power diversity-dependence on the extinction rate; to be
fed as argument \code{dd_model} to \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
}
\details{
\deqn{\lambda(N) = \lambda_{0}}
\deqn{\mu(N) = \mu_{0} \times N^{\frac{log\Big(\frac{\lambda_0 - \mu_0}{\lambda_0}+1\Big)}{log(K)}}}
}
\author{
Theo Pannetier
}
