% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_cx.R
\name{dd_model_cx}
\alias{dd_model_cx}
\title{Diversity-dependent model with constant-rate (c) speciation and
exponential (x) dependence on extinction}
\usage{
dd_model_cx()
}
\description{
A list specifying a DD model with constant-rate
speciation and exponential diversity-dependence on the extinction rate;
to be fed as argument \code{dd_model} to \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
}
\details{
\deqn{\lambda(N) = \lambda_{0}}
\deqn{\mu(N) = \mu_{0}(\frac{\lambda_{0}}{\mu_{0}})^{\frac{N}{K}}}
}
\author{
Theo Pannetier
}
