% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_ll.R, R/dd_model_lx.R
\name{dd_model_ll}
\alias{dd_model_ll}
\title{Diversity-dependent model with linear (l) dependence on speciation and
linear (l) dependence on extinction}
\usage{
dd_model_ll()

dd_model_ll()
}
\description{
A list specifying a DD model with linear diversity-dependence on the
speciation rate and extinction rate; to be fed as argument
\code{dd_model} to \code{\link[=fit_dd_model]{fit_dd_model()}}.

A list specifying a DD model with linear diversity-dependence on the
speciation rate and exponential diversity-dependence on the extinction rate;
to be fed as argument \code{dd_model} to \code{\link[=fit_dd_model]{fit_dd_model()}}.
}
\details{
\deqn{\lambda(N) = \lambda_{0} - (1 - \alpha)(\lambda_{0} - \mu_{0}) \frac{N}{K}}
\deqn{\mu(N) = \mu_{0} + \alpha(\lambda_{0} - \mu_{0}) \frac{N}{K}}

\deqn{\lambda(N) = \lambda_{0} - (1 - \alpha)(\lambda_{0} - \mu_{0}) \frac{N}{K}}
\deqn{\mu(N) = \mu_{0}((1 - \alpha) + \alpha \frac{\lambda_{0}}{\mu_{0}})^{\frac{N}{K}}}
}
\author{
Theo Pannetier

Theo Pannetier
}
