% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_xc2.R
\name{dd_model_xc2}
\alias{dd_model_xc2}
\title{Diversity-dependent model with exponential (x) dependence on speciation and
constant (c) extinction – alternative formulation}
\usage{
dd_model_xc2()
}
\description{
A list specifying a DD model with exponential diversity-dependence on the
speciation rate and constant-rate extinction; to be fed as argument
\code{dd_model} to \code{\link[=fit_dd_model]{fit_dd_model()}}. This second formulation of the model
corresponds to the exponential model found in \code{DDD} (\code{ddmodel = 2})
}
\details{
\deqn{\lambda(N) = \lambda_{0} \times N^{-\frac{log(\frac{\lambda_{0}}{\mu_{0}})}{log(K)}} \\ \mu(N) = \mu_{0}}
}
\author{
Theo Pannetier
}
