% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_xx2.R
\name{dd_model_xx2}
\alias{dd_model_xx2}
\title{Diversity-dependent model with exponential (x) dependence on speciation and
exponential (x) dependence on extinction – alternative formulation}
\usage{
dd_model_xx2()
}
\description{
A list specifying a DD model with exponential diversity-dependence on both the
speciation rate and extinction rate; to be fed as argument \code{dd_model} to
\code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}. This formulation corresponds to the exponential model
found in \code{DDD} (\code{ddmodel = 2}) and elsewhere (\code{BAMM}).
}
\details{
\deqn{\lambda(N) = \lambda_{0} \times N^{-\frac{log\Big(\frac{\lambda_0}{\alpha(\lambda_0 - \mu_0) + \mu_0}\Big)}{log(K)}}}
\deqn{\mu(N) = \mu_{0} \times N^{\frac{log\Big(\alpha\frac{\lambda_0 - \mu_0}{\lambda_0}+1\Big)}{log(K)}}}
}
\author{
Theo Pannetier
}
