% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars.R
\name{default_pars}
\alias{default_pars}
\alias{default_sigma_comp}
\alias{default_carr_cap_pars}
\alias{default_growth_rate}
\title{Default parameter values}
\usage{
default_sigma_comp()

default_carr_cap_pars()

default_growth_rate()
}
\description{
Return default values of the simulation parameters, i.e. the default settings
used in Pontarp et al. (2012).
}
\author{
Theo Pannetier
}
