% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_init_params_dd_ml.R
\name{draw_init_params_dd_ml}
\alias{draw_init_params_dd_ml}
\title{Draw a set of initial parameter values for a DD model}
\usage{
draw_init_params_dd_ml(phylos, nb_sets)
}
\arguments{
\item{phylos}{the list of `phylo`` class objects on which the DD model is going
to be fit.}

\item{nb_sets}{integer, the number of sets desired.}
}
\value{
a list of length-3 vectors containing initial values for \code{lambda_0},
\code{mu_0} and \code{K}.
}
\description{
Draw a set of initial parameter values for a DD model
}
\author{
Theo Pannetier
}
