% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_nb_offspring.R
\name{draw_nb_offspring}
\alias{draw_nb_offspring}
\title{Draw a number of offspring given a fitness value}
\usage{
draw_nb_offspring(fitness)
}
\arguments{
\item{fitness}{numeric \verb{>= 0}, a vector of fitness values.}
}
\description{
The number of offspring is drawn in a Poisson distribution in
\code{rpois(lambda = fitness)}. Vectorized.
}
\seealso{
get_fitness
}
\author{
Theo Pannetier
}
