% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pop_next_gen.R
\name{draw_pop_next_gen}
\alias{draw_pop_next_gen}
\title{Draw a new population from the current one}
\usage{
draw_pop_next_gen(
  pop,
  growth_rate = default_growth_rate(),
  comp_width = default_comp_width(),
  trait_opt = default_trait_opt(),
  carr_cap_opt = default_carr_cap_opt(),
  carr_cap_width = default_carr_cap_width(),
  prob_mutation = default_prob_mutation(),
  mutation_sd = default_mutation_sd()
)
}
\arguments{
\item{pop}{a tibble with one row per individual in the population and two
columns:
\itemize{
\item \code{z} contains (numeric) traits values
\item \code{species} contains the species names (characters)
}}

\item{growth_rate}{numeric \eqn{>= 0}, the baseline growth rate in the
absence of competition.}

\item{comp_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{\alpha}}. Controls
the intensity of competition given a distance between two trait values.}

\item{trait_opt}{numeric. Optimal trait value \eqn{z_{opt}}, such that
\code{get_carr_cap(trait_opt) = carr_cap_opt}.}

\item{carr_cap_opt}{numeric. Maximum carrying capacity at \eqn{z = z_{opt}}.
Note that this corresponds to the maximum number of \strong{competitors} for that
trait value, as the focal individual does not contribute to the carrying
capacity (see also \code{\link{get_fitness}}).}

\item{carr_cap_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{K}}. Controls
how fast the carrying capacity decays as the distance between a \code{trait}
and \code{trait_opt} increases.
\eqn{\sigma^{2}_{K}}}

\item{prob_mutation}{numeric between 0 and 1, the probability that any new
individual is sampled with a mutation.}

\item{mutation_sd}{numeric \eqn{>= 0}, the standard deviation of the normal
distrbution in which the mutations are drawn.}
}
\description{
Based on the current population traits, compute the fitness, draw
offspring, apply mutations and speciation events were relevant.
}
\author{
Théo Pannetier
}
