% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_species_name.R
\name{draw_species_name}
\alias{draw_species_name}
\title{Draw a made-up name for a species}
\usage{
draw_species_name(genus_name = "Haggis")
}
\arguments{
\item{genus_name}{character, a genus name.}
}
\description{
Draw a random species name and combines it with the supplied genus.
}
\author{
Théo Pannetier
}
