% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dd_rates.R
\name{estimate_dd_rates}
\alias{estimate_dd_rates}
\title{Estimate diversity-dependent rates from phylogenies}
\usage{
estimate_dd_rates(multi_phylo)
}
\arguments{
\item{multi_phylo}{a list of \code{phylo} objects, as typically created with
\code{ape}.}
}
\value{
a \code{tibble} with a row for every possible number of species and 4
columns:
\itemize{
\item \code{N}, level of diversity, nb of species
\item \code{nb_events}, the number of replicates for that level of diversity; that is,
the number of events corresponding to that state of diversity, across all
phylogenies in \code{multi_phylo}.
\item \code{speciation_rate} the reconstructed, average, per-capita speciation rate
for that level of diversity
\item \code{extinction_rate} the reconstructed, average, per-capita extinction rate
for that level of diversity
}


}
\description{
For a list of phylogenies, reconstruct the \emph{average, per-capita} rates of
speciation and extinction corresponding to each level of species diversity.
}
\author{
Théo Pannetier
}
