% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_trait_gaps.R
\name{find_trait_gaps}
\alias{find_trait_gaps}
\title{Find gaps in trait values}
\usage{
find_trait_gaps(traits)
}
\arguments{
\item{traits}{a numeric vector, \strong{ordered} by ascending values.}
}
\description{
Runs through an ordered vector of trait values, returns the positions of gaps
between consecutive values.
}
\author{
Théo Pannetier
}
