% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dd_model.R
\name{fit_dd_model}
\alias{fit_dd_model}
\title{Fit a diversity-dependent diversification model to phylogenies}
\usage{
fit_dd_model(phylos, init_params, speciation_func, extinction_func)
}
\arguments{
\item{phylos}{a list of \code{phylo} objects, typically the output of
\code{\link[=sim_to_phylo]{sim_to_phylo()}}. The phylogenies must be complete, i.e extinct species must
be included.}

\item{init_params}{named vector, initial values of the parameters to optimise.
Names must match the parameters called in \code{speciation_func} and
\code{extinction_func}.}

\item{speciation_func}{a function describing how the rate of speciation varies
with the number of species \code{N}. Must take arguments \code{params} and \code{N}. Inside
the function, parameters must be referred as \code{params$param_name}.}

\item{extinction_func}{a function describing how the rate of extinction varies
with the number of species \code{N}. Must take arguments \code{params} and \code{N}. Inside
the function, parameters must be referred as \code{params$param_name}.}
}
\value{
a one-row table with maximum likelihood estimates for the values of
the parameters, the maximum likelihood and a convergence code (see
\code{\link[subplex:subplex]{subplex::subplex()}} for the meaning of this code)
}
\description{
Fits the parameters of a specified diversity-dependent model to  a set of
full phylogenies produced by \code{comrad} simulations
}
\author{
Theo Pannetier
}
