% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dd_model_without_fossil.R
\name{fit_dd_model_without_fossil}
\alias{fit_dd_model_without_fossil}
\title{Fit a diversity-dependent diversification model to phylogenies without fossil lineages}
\usage{
fit_dd_model_without_fossil(
  branching_times,
  init_params,
  dd_model = dd_model_lc(),
  num_cycles = Inf,
  methode = "odeint::runge_kutta_cash_karp54",
  maxiter = 1000 * round((1.25)^length(init_params)),
  verbose = FALSE
)
}
\arguments{
\item{branching_times}{a vector containing the branching times of a single, ultrametric
(all tips reach the present) phylogenetic tree. Tip: from the full \code{phylo}, do
\code{phylo \%>\% ape::drop.fossil() \%>\% ape::branching.times()}.}

\item{init_params}{named vector, initial values of the parameters to optimise.
The names and number of parameters must match those specified in
\code{dd_model$params_check}.}

\item{dd_model}{a list with five named elements that together specify the
diversity-dependent model:
\itemize{
\item \code{name} a two-letter code, the name of the model. First letter specifies the
form of the speciation function, second letter the form of the extinction
function: "l" for "linear", "x" for exponential, "c" for constant.
\item \code{speciation_func}, a function specifying the diversity-dependent speciation
rate. Must take arguments \code{params} and \code{N}.
\item \code{extinction_func}, a function specifying the diversity-dependent extinction
rate. Must take arguments \code{params} and \code{N}.
\item \code{constraints} a list of conditions that parameter values must satisfy. Each
element is a function that takes arguments \code{params} and \code{...}, and returns
\code{TRUE} if the constraint is satisfied, \code{FALSE} if it isn't.
\item \code{params_check} a function that controls the format of \code{params}. Returns an
error if the elements of \code{params} are named differently from what is expected
or if the length differs from the expectation.
}

\code{comrad} contains several \code{dd_model} functions, see for example
\code{\link[=dd_model_lc]{dd_model_lc()}}.}

\item{num_cycles}{passed to \code{\link[DDD:optimizer]{DDD::optimizer()}}, number of cycles of
optimisation. Next cycle starts from the last vertices of the previous cycle.}

\item{methode}{passed to \code{\link[DDD:optimizer]{DDD::optimizer()}}, a character specifying the
methode to solve the DD master system, either numerically ("analytical") or
numerically ("odeint::runge_kutta_cash_karp54" or another function from the
\code{odeint} library)}

\item{maxiter}{integer, passed to \code{DDD}'s optimizer.
The maximum number of iterations for a single cycle of optimization.}

\item{verbose}{logical, should the output of every optimisation iteration be
printed to console?}
}
\value{
a one-row table with the model name, initial values of the parameters,
maximum likelihood estimates and the maximum likelihood
}
\description{
Wrapper around \code{\link[DDD:dd_ML]{DDD::dd_ML()}}. Fits the parameters of a specified
diversity-dependent model the reconstructed phylogeny of a single tree
produced by \code{comrad} simulations.
}
\author{
Theo Pannetier
}
