% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{Get the fitness of an individual}
\usage{
fitness(
  trait_ind,
  traits_pop,
  growth_rate = default_growth_rate(),
  sigma_comp = default_sigma_comp(),
  carr_cap_pars = default_carr_cap_pars()
)
}
\arguments{
\item{trait_ind}{numeric. Trait value \eqn{z} of a focal individual.}

\item{traits_pop}{numeric vector of variable length, the trait values of all
competitors in the population, excluding the focal individual.}

\item{growth_rate}{numeric \eqn{>= 0}, the baseline growth rate in the
absence of competition.}

\item{sigma_comp}{numeric > 0. Controls the competition intensity, i.e. the
strength of negative feedbacks the focal indiviudal experiences for a given
trait distance with the competitor.}

\item{carr_cap_pars}{parameter vector with three elements: \enumerate{
   \item \code{trait_opt} numeric. Optimal trait value \eqn{z_{opt}}, at which
\eqn{K = K_{0}}
   \item carr_cap_opt numeric. Maximum carrying capacity at \eqn{z = z_{opt}}.
Note that this corresponds to the maximum number of **competitors** for that
trait value, as the focal individual does not contribute to the carrying
capacity (see also \code{\link{fitness}}).
   \item carr_cap_var numeric. Variance of the carrying capacity
\eqn{\sigma^{2}_{K}}
}}
}
\description{
Computes the fitness of an individual based on its trait value and the trait
values of other individuals in the population. The fitness here corresponds
to the average number of offspring an individual can get in a generation, as
sampled in a Poisson distribution.
}
\details{
The equation is a per-capita version of the Ricker model:
\deqn{G(z_i) = exp(r(1 - N_{eff} / K(z_i, z_opt)))}
where \eqn{N_{eff}} is the effective population size, i.e. sum of competitive
effects experienced by the focal individual for a given trait value:
\deqn{N_{eff} = sum_j(\alpha(z_i, z_j))}
}
\author{
Theo Pannetier
}
