% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_coeff_pair.R
\name{get_comp_coeff_pair}
\alias{get_comp_coeff_pair}
\title{Get the competition coefficient between two individuals}
\usage{
get_comp_coeff_pair(
  trait_ind_one,
  trait_ind_two,
  comp_width = default_comp_width()
)
}
\arguments{
\item{trait_ind_one}{numeric, trait value of a focal individual.}

\item{trait_ind_two}{numeric, trait value of a competign individual.}

\item{comp_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{\alpha}}. Controls
the intensity of competition given a distance between two trait values.}
}
\description{
Alternative to \code{\link[=get_comp_coeff_pop]{get_comp_coeff_pop()}} on a single pair of individuals,
intended to save time on the number of computations. Note that the method of
computation is different between the two functions, so \code{\link[=get_comp_coeff_pop]{get_comp_coeff_pop()}}
is \emph{not} a vectorized version of \code{get_comp_coeff_pair()}.
}
\author{
Théo Pannetier
}
