% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_coeffs.R
\name{get_comp_coeffs}
\alias{get_comp_coeffs}
\title{Get the competition coefficient between two individuals}
\usage{
get_comp_coeffs(trait_ind, traits_pop, comp_width = default_comp_width())
}
\arguments{
\item{trait_ind}{numeric. Trait value \eqn{z} of a focal individual.}

\item{traits_pop}{numeric vector of variable length, the trait values of
every individual in the population.}

\item{comp_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{\alpha}}. Controls
the intensity of competition given a distance between two trait values.}
}
\value{
a numeric vector containing the competitive effect caused by each
individual in the population on the focal individual.
}
\description{
Returns the competition coefficients \eqn{\alpha(zi, z_j)} between a focal
individual \eqn{i}, and each individual \eqn{j} in the population, including
the focal individual itself.
}
\details{
The equation is reported from equation (3) in Pontarp et al. (2012).
\deqn{\alpha(z_i, z_j) = exp(-(z_i - z_j) ^ 2 / (2 * \sigma^2_\alpha))}
}
\author{
Theo Pannetier
}
