% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eff_pop_sizes.R
\name{get_eff_pop_sizes}
\alias{get_eff_pop_sizes}
\title{Compute effective population sizes for a population}
\usage{
get_eff_pop_sizes(traits_pop, sigma_comp = default_sigma_comp())
}
\arguments{
\item{traits_pop}{numeric vector of variable length, the trait values of all
competitors in the population.}

\item{sigma_comp}{numeric > 0. Controls the competition intensity, i.e. the
strength of negative feedbacks the focal indiviudal experiences for a given
trait distance with the competitor.}
}
\description{
Computes \eqn{n_eff}, the effective population size experienced by an
individual based on its trait value and the rest of the population.
Vectorized, can take either individual or population trait values as input.
}
\author{
Theo Pannetier
}
