% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitness_landscape.R
\name{get_fitness_landscape}
\alias{get_fitness_landscape}
\title{Compute the fitness along a range of trait values, given a community}
\usage{
get_fitness_landscape(
  z_seq,
  traits_comm,
  growth_rate = default_growth_rate(),
  comp_width = default_comp_width(),
  carr_cap_opt = default_carr_cap_opt(),
  carr_cap_width = default_carr_cap_width()
)
}
\arguments{
\item{z_seq}{numeric vector, the sequence of trait values along which the
fitness landscape should be computed.}

\item{traits_comm}{numeric vector, the trait values of all individuals in the
community for which the fitness landscape is to be computed}

\item{growth_rate}{numeric \verb{>= 0}, the baseline growth rate. Generations
being discrete, high values will cause chaos.}

\item{comp_width}{numeric \verb{>= 0}. Width of the competition kernel.}

\item{carr_cap_opt}{numeric. Carrying capacity at the optimal trait value.}

\item{carr_cap_width}{numeric \verb{>= 0}.
Width of the carrying capacity kernel.}
}
\value{
a \code{tibble} with \code{z_seq} rows and columns \code{z}, \code{fitness}.
}
\description{
Computes the fitness values at every point in \code{z_seq}, given the community
defined by \code{traits_comm}. This is different from \code{\link[=get_fitness]{get_fitness()}},
where the fitness is only computed for the trait values of individuals in the
community.
}
\author{
Théo Pannetier
}
