% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_n_eff_seq}
\alias{get_n_eff_seq}
\title{Compute the effective population size at a range of values}
\arguments{
\item{z_seq}{numeric vector, the set of trait values for which we want to know
\code{n_eff}, given \code{z_pop}}

\item{z_pop}{numeric vector, the trait values of all individuals in the population.}

\item{competition_sd}{numeric \verb{>= 0}. Width of the competition kernel.}
}
\description{
Given the trait values of a population \code{z_pop}, calculate the effective
population sizes \code{n_eff} experienced at a set of positions \code{z_seq}
on the trait axis.
}
\details{
\code{n_eff} sums the competitive effects an individual receives from
every individual in the community, including the individual itself. It is
called effective population size because it is the size of the population
that is relevant for competition.
}
\author{
Théo Pannetier
}
