% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_eff_shortcut.R
\name{get_n_eff_shortcut}
\alias{get_n_eff_shortcut}
\title{Computes the effective population size with less computations}
\usage{
get_n_eff_shortcut(traits_pop, comp_width = default_comp_width())
}
\arguments{
\item{traits_pop}{numeric vector of variable length, the trait values of
every individual in the population.}

\item{comp_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{\alpha}}. Controls
the intensity of competition given a distance between two trait values.}
}
\description{
This function is intended to do the same job as \code{\link[=get_n_eff]{get_n_eff()}}, but with less
computations.
}
\details{
For each individual, \code{\link[=get_n_eff]{get_n_eff()}} computes the sum of its
interactions (\code{\link[=get_comp_coeff_pair]{get_comp_coeff_pair()}}) with every other individual in the
population. Because these interactions are symmetric
(i.e. \code{comp_coeff(x, y) == comp_coeff(y, x)}), \code{\link[=get_n_eff]{get_n_eff()}} does
more calculations than is actually necessary. \code{get_n_eff_shortcut()} here is
its lazy twin, that computes the competition coefficient once for each unique
pair of individuals.
}
\author{
Theo Pannetier
}
