% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comm_bubbles.R
\name{plot_comm_bubbles}
\alias{plot_comm_bubbles}
\title{Plot population densities and trait distribution over time, with bubbles}
\usage{
plot_comm_bubbles(comrad_tbl, generation_range = c(0, Inf))
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}

\item{generation_range}{numeric vector with length 2, supplying the first and
last generation to plot from the dataset. Defaults to every generation in the
table}
}
\description{
Produces a plot of trait evolution where individuals are binned as bubbles
which colour indicate the species identity and the size the number of
individuals in the bin. Equivalent to \code{\link[=plot_comm_trait_evolution]{plot_comm_trait_evolution()}}, plus
species identities.
}
\author{
Théo Pannetier
}
