% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comm_size.R
\name{plot_comm_size}
\alias{plot_comm_size}
\title{Plot community size over generations}
\usage{
plot_comm_size(comrad_tbl)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}
}
\description{
Plots the number of individuals in the community over time.
}
\author{
Théo Pannetier
}
