% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generation_traits.R
\name{plot_generation_traits}
\alias{plot_generation_traits}
\title{Plot trait distribution in a generation}
\usage{
plot_generation_traits(path_to_file, generation, binwidth = 0.02)
}
\arguments{
\item{path_to_file}{character, path to a \code{.csv} file as produced by
\code{\link[=run_simulation]{run_simulation()}}}

\item{generation}{numeric, the index of the generation to plot}

\item{binwidth}{numeric, the width used for binning.}
}
\description{
Reads a simulation data table and plots a histogram of the distribution of
trait values in a single generation.
}
\author{
Théo Pannetier
}
