% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_population_size.R
\name{plot_population_size}
\alias{plot_population_size}
\title{Plot population size over generations}
\usage{
plot_population_size(path_to_file)
}
\arguments{
\item{path_to_file}{character, path to a \code{.csv} file as produced by
\code{\link[=run_simulation]{run_simulation()}}}
}
\description{
Reads a simulation data table and plots the number of individuals in the
population over time
}
\author{
Théo Pannetier
}
