% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_test.R
\name{robustness_test}
\alias{robustness_test}
\title{Run a robustness test as a measure of absolute fit}
\usage{
robustness_test(ltts_empirical, ltts_bootstrap)
}
\arguments{
\item{ltts_empirical}{a list of \code{tibble} with columns \code{time} and \code{N} (as produced by \code{\link[=get_ltt_tbl]{get_ltt_tbl()}})
containing the LTTs of the set of empirical trees.}

\item{ltts_bootstrap}{a list of \code{tibble} with columns \code{time} and \code{N} (as produced by \code{\link[=get_ltt_tbl]{get_ltt_tbl()}})
containing the LTTs of the set of bootstrap trees.}
}
\value{
a list with four elements:
\itemize{
\item \code{dnltts_empirical} vector with the delta nLTT between every tree in the empirical set and the average bootstrap LTT
\item \code{dnltts_bootstrap} vector with the delta nLTT between every tree in the bootsrap set and the average bootstrap LTT
\item \code{p_95} a measure akin a p-value, the proportion of empirical trees with a delta nLTT higher than the 95th percentile of the distribution delta nLTT of the bootstrap test.
Based on the $p_95$ of Santos Neves, Lambert et al. (2021).
\item \code{ks} output of a two-sample, one-sided (empirical > bootstrap) Kolmogorv-Smirnov test on the distribution of the delta nLTT of the empirical and bootsrap test.
}
}
\description{
This function will run a test that is meant as a measure of the absolute fit
of a DD model to a set of comrad trees. The test uses the delta nLTT statistic
(Janzen et al. 2015) as a measure of error between the branching pattern expected under a
DD model and that observed in a set of comrad trees. Inspired from the bootstrap tests
of Louca & Pennell (2021) and Santos Neves, Lambert et al. (2021). Deterministic LTT of the DD
model estimated as the average LTT of the bootstrap set.
}
\details{
Here, "empirical trees" is meant as a set of trees to which the fit of a DD model is to be assessed,
while "bootstrap trees" are a set of trees simulated under the DD model, with maximum likelihood estimates for the
set of empirical trees.
}
\author{
Theo Pannetier
}
