% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(
  init_pop,
  output_path,
  nb_generations,
  sampling_frequency = set_sampling_frequency(nb_generations),
  seed = default_seed(),
  growth_rate = default_growth_rate(),
  comp_width = default_comp_width(),
  trait_opt = default_trait_opt(),
  carr_cap_opt = default_carr_cap_opt(),
  carr_cap_width = default_carr_cap_width(),
  mutation_sd = default_mutation_sd(),
  fitness_func = fitness_func_positive_logistic
)
}
\arguments{
\item{init_pop}{numeric vector, the trait values of the initial individuals.}

\item{output_path}{character, path to save the output file.}

\item{nb_generations}{integer, the number of generations to run during the
simulation.}

\item{sampling_frequency}{numeric \code{> 0}, the frequency at which the
population is saved in the output.}

\item{seed}{numeric \code{> 0}, the integer seed to set for the random number
generator.}

\item{growth_rate}{numeric \eqn{>= 0}, the baseline growth rate in the
absence of competition.}

\item{comp_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{\alpha}}. Controls
the intensity of competition given a distance between two trait values.}

\item{trait_opt}{numeric. Optimal trait value \eqn{z_{opt}}, such that
\code{get_carr_cap(trait_opt) = carr_cap_opt}.}

\item{carr_cap_opt}{numeric. Maximum carrying capacity at \eqn{z = z_{opt}}.
Note that this corresponds to the maximum number of **competitors** for that
trait value, as the focal individual does not contribute to the carrying
capacity (see also \code{\link{get_fitness}}).}

\item{carr_cap_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{K}}. Controls
 how fast the carrying capacity decays as the distance between a \code{trait}
 and \code{trait_opt} increases.
\eqn{\sigma^{2}_{K}}}

\item{mutation_sd}{numeric \eqn{>= 0}, the standard deviation of the normal
distrbution in which the mutations are drawn.}

\item{fitness_func}{object. The function to use to compute fitness.}
}
\description{
Run the competitive radiation simulation.
}
\details{
Output is registered in a .csv file with the following structure:
| t | z | runtime |
| --- | --- |--- |
| --- | --- |--- |
where each line is an individual (the entire population is recorded).
\code{t} denotes the generation counter, \code{z} is a trait value, and
\code{runtime} is the processing time elapsed during the current generation.
In addition, the output table is preceded by metadata. Skip the 14 first
lines to get the table proper.
}
\author{
Theo Pannetier
}
