% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sampling_frequency.R
\name{set_sampling_frequency}
\alias{set_sampling_frequency}
\title{Set default sampling frequency}
\usage{
set_sampling_frequency(nb_gens)
}
\arguments{
\item{nb_gens}{integer, the number of generations to run during the
simulation.}
}
\description{
Returns an appropriate sampling frequency based on the order of magnitude of
\code{nb_gens}.
}
\details{
The sampling frequency is set as follows:
\itemize{
\item if \code{nb_gens < 1000}, sampling frequency is set to 1
(every generation).
\item if \code{nb_gens >= 1000}, sampling frequency is set to twice
the order of magnitude of \code{nb_gens} - 2. E.g. for
\code{nb_gens >= 1000}, the community is sampled every 20
generations, every 200 generations for \code{nb_gens >= 10000},
etc.
}
}
\author{
Theo Pannetier
}
