% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_brownian_mutation.R
\name{simulate_brownian_mutation}
\alias{simulate_brownian_mutation}
\title{Simulate a population undergoing brownian mutations}
\usage{
simulate_brownian_mutation(
  pop_size = 1000,
  nb_gen = 1000,
  mutation_sd = default_mutation_sd()
)
}
\arguments{
\item{pop_size}{integer, size of the population to simulate.}

\item{nb_gen}{integer, number of generations to run the simulation for.}

\item{mutation_sd}{numeric \eqn{>= 0}, the standard deviation of the normal
distrbution in which the mutations are drawn.}
}
\description{
This function serves to test the mutation rate in the main simulation. Runs
and plots trait values undergoing brownian motion under the specified
mutation rate.
}
\author{
Théo Pannetier
}
