% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targs.R
\name{targs}
\alias{targs}
\alias{targ_num}
\alias{targ_pos}
\alias{targ_prop}
\alias{targ_not_zero}
\alias{targ_char}
\alias{targ_log}
\alias{targ_length}
\title{Assert a function argument is of correct type}
\usage{
targ_num(arg_val, arg_name)

targ_pos(arg_val, arg_name)

targ_prop(arg_val, arg_name)

targ_not_zero(arg_val, arg_name)

targ_char(arg_val, arg_name)

targ_log(arg_val, arg_name)

targ_length(arg_val, arg_name, correct_length)
}
\arguments{
\item{arg_val}{value of the asserted argument.}

\item{arg_name}{name of the asserted argument to be passed to an error message}

\item{correct_length}{numeric, the length the argument should have.}
}
\description{
\itemize{
 \item \code{targ_num()} asserts that the argument is a numeric value.
 \item \code{targ_pos()} asserts that the argument is a numeric value and
 positive.
 \item \code{targ_not_zero()} asserts that the argument is a numeric value
  different from 0
 \item \code{targ_char()} asserts that the argument is a character.
 \item \code{targ_log()} asserts that the argument is a logical.
 \item \code{targ_length()} asserts that the argument has the correct length.
}
}
\author{
Theo Pannetier
}
