% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiting_times.R
\name{waiting_times}
\alias{waiting_times}
\title{Extract times to events from a phylogeny}
\usage{
waiting_times(phylo, pool_events = TRUE)
}
\arguments{
\item{phylo}{an object of the class \code{phylo} as introduced in
\link[ape:read.tree]{ape}. The phylogeny must start with the crown node
(not stem), and be binary (no hard polytomies).#'}

\item{pool_events}{logical, do you want the waiting time to next event
(regardless of speciation, extinction), or next event of same type?}
}
\value{
a \code{tibble} with a row per event and four columns:
\itemize{
\item \code{time}, time of an event
\item \code{N}, number of species in the tree at that time
\item \code{next_event}, type of the next event; either \code{"speciation"} or
\code{"extinction"}
\item \code{waiting_time}, time to the next event
}
}
\description{
Turns the phylogeny into a lineage-through-time table, and for which
event computes the time to next event.
}
\author{
Théo Pannetier
}
