% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carr_cap.R
\name{carr_cap}
\alias{carr_cap}
\title{Get the carrying capacity for a given trait value}
\usage{
carr_cap(
  trait_ind,
  trait_opt = default_carr_cap_pars()[1],
  carr_cap_opt = default_carr_cap_pars()[2],
  carr_cap_var = default_carr_cap_pars()[3]
)
}
\arguments{
\item{trait_ind}{numeric. Trait value \eqn{z} of a focal individual.}

\item{trait_opt}{numeric. Optimal trait value \eqn{z_{opt}}, at which
\eqn{K = K_{0}}}

\item{carr_cap_opt}{numeric. Maximum carrying capacity at \eqn{z = z_{opt}}.
Note that this corresponds to the maximum number of **competitors** for that
trait value, as the focal individual does not contribute to the carrying
capacity (see also \code{\link{get_fitness}}).}

\item{carr_cap_var}{numeric. Variance of the carrying capacity
\eqn{\sigma^{2}_{K}}}
}
\description{
Computes the carrying capacity experienced by an individual with trait value
\code{trait}.
}
\author{
Theo Pannetier
}
