% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  trait_ind,
  traits_pop,
  comp_with,
  trait_opt,
  carr_cap_opt,
  carr_cap_width,
  growth_rate,
  carr_cap_pars
)
}
\arguments{
\item{trait_ind}{numeric. Trait value \eqn{z} of a focal individual.}

\item{traits_pop}{numeric vector of variable length, the trait values of
every individual in the population.}

\item{trait_opt}{numeric. Optimal trait value \eqn{z_{opt}}, such that
\code{get_carr_cap(trait_opt) = carr_cap_opt}.}

\item{carr_cap_opt}{numeric. Maximum carrying capacity at \eqn{z = z_{opt}}.
Note that this corresponds to the maximum number of **competitors** for that
trait value, as the focal individual does not contribute to the carrying
capacity (see also \code{\link{get_fitness}}).}

\item{carr_cap_width}{numeric ≥ 0. \eqn{\sigma^{2}_{K}}. Controls how fast
the carrying capacity decays as the distance between a \code{trait} and
\code{trait_opt} increases.
\eqn{\sigma^{2}_{K}}}

\item{growth_rate}{numeric \eqn{>= 0}, the baseline growth rate in the
absence of competition.}

\item{carr_cap_pars}{parameter vector with three elements: \enumerate{
   \item \code{trait_opt} numeric. Optimal trait value \eqn{z_{opt}}, at
   which \eqn{K = K_{0}}.
   \item carr_cap_opt numeric. Maximum carrying capacity at
   \eqn{z = z_{opt}}.
Note that this corresponds to the maximum number of **competitors** for that
trait value, as the focal individual does not contribute to the carrying
capacity (see also \code{\link{get_fitness}}).
   \item carr_cap_with numeric. Variance of the carrying capacity
\eqn{\sigma^{2}_{K}}
}}

\item{comp_width}{numeric ≥ 0. \eqn{\sigma^{2}_{\alpha}}. Controls the
intensity of competition given a distance between two trait values.}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\author{
Theo Pannetier, based on skeleton stolen from Richel J.C. Bilderbeek.
}
