% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation for main parameters of the model}
\usage{
default_params_doc(
  pop,
  trait_ind,
  traits_pop,
  comp_width,
  trait_opt,
  carr_cap_opt,
  carr_cap_width,
  growth_rate,
  prob_mutation,
  mutation_sd,
  fitness,
  nb_generations,
  fitness_func,
  sim_tbl
)
}
\arguments{
\item{pop}{a tibble with one row per individual in the population and two
columns:
\itemize{
\item \code{z} contains (numeric) traits values
\item \code{species} contains the species names (characters)
}}

\item{trait_ind}{numeric. Trait value \eqn{z} of a focal individual.}

\item{traits_pop}{numeric vector of variable length, the trait values of
every individual in the population.}

\item{comp_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{\alpha}}. Controls
the intensity of competition given a distance between two trait values.}

\item{trait_opt}{numeric. Optimal trait value \eqn{z_{opt}}, such that
\code{get_carr_cap(trait_opt) = carr_cap_opt}.}

\item{carr_cap_opt}{numeric. Maximum carrying capacity at \eqn{z = z_{opt}}.
Note that this corresponds to the maximum number of \strong{competitors} for that
trait value, as the focal individual does not contribute to the carrying
capacity (see also \code{\link{get_fitness}}).}

\item{carr_cap_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{K}}. Controls
how fast the carrying capacity decays as the distance between a \code{trait}
and \code{trait_opt} increases.
\eqn{\sigma^{2}_{K}}}

\item{growth_rate}{numeric \eqn{>= 0}, the baseline growth rate in the
absence of competition.}

\item{prob_mutation}{numeric between 0 and 1, the probability that any new
individual is sampled with a mutation.}

\item{mutation_sd}{numeric \eqn{>= 0}, the standard deviation of the normal
distrbution in which the mutations are drawn.}

\item{fitness}{numeric \eqn{>= 0} vector, containing the fitness values for
each individual as computed by \code{\link{get_fitness}}}

\item{nb_generations}{integer, the number of generations to run during the
simulation.}

\item{fitness_func, }{object. The function to use to compute fitness.}

\item{sim_tbl}{a tibble containing the output of a comrad simulation, as
produced by \code{\link[=run_simulation]{run_simulation()}} and read by \code{\link[=read_sim_tbl]{read_sim_tbl()}}.}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\author{
Theo Pannetier, based on skeleton stolen from Richel J.C. Bilderbeek.
}
