% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation for main parameters of the model}
\usage{
default_params_doc(
  comm,
  trait_ind,
  traits_comm,
  comp_width,
  trait_opt,
  carr_cap_opt,
  carr_cap_width,
  growth_rate,
  prob_mutation,
  mutation_sd,
  trait_gap,
  nb_gens,
  fitness_func,
  comrad_tbl
)
}
\arguments{
\item{comm}{a tibble with one row per individual in the community and three
columns:
\itemize{
\item \code{z} contains (numeric) traits values
\item \code{species} contains species names (characters)
\item \code{ancestral_species} contains acnestral species names (characters)
}}

\item{trait_ind}{numeric. Trait value \code{z} of a focal individual.}

\item{traits_comm}{numeric vector of variable length, the trait values of
every individual in the community.}

\item{comp_width}{numeric \verb{>= 0}. Width of the competition kernel.}

\item{trait_opt}{numeric. The optimal trait value.
\code{get_carr_cap(trait_opt) = carr_cap_opt}.}

\item{carr_cap_opt}{numeric. Carrying capacity at the optimal trait value.}

\item{carr_cap_width}{numeric \verb{>= 0}.
Width of the carrying capacity kernel.}

\item{growth_rate}{numeric \verb{>= 0}, the baseline growth rate. Generations
being discrete, high values will cause chaos.}

\item{prob_mutation}{numeric between 0 and 1, the probability that any new
individual is sampled with a mutation.}

\item{mutation_sd}{numeric \verb{>= 0}, the standard deviation of the normal
distrbution from which mutations are drawn.}

\item{trait_gap}{numeric, the width of a gap triggering speciation.}

\item{nb_gens}{integer, how many generations should the simulation be run
for?}

\item{fitness_func, }{the name of the function used to compute fitness.}

\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}} or read by \code{\link[=read_comrad_tbl]{read_comrad_tbl()}}.}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\author{
Theo Pannetier, based on skeleton stolen from Richel J.C. Bilderbeek.
}
