% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars.R
\name{default_pars}
\alias{default_pars}
\alias{default_comp_width}
\alias{default_trait_opt}
\alias{default_carr_cap_opt}
\alias{default_carr_cap_width}
\alias{default_growth_rate}
\alias{default_mutation_sd}
\alias{default_init_pop}
\alias{default_seed}
\title{Default parameter values}
\usage{
default_comp_width()

default_trait_opt()

default_carr_cap_opt()

default_carr_cap_width()

default_growth_rate()

default_mutation_sd()

default_init_pop()

default_seed()
}
\description{
Return default values of the simulation parameters, i.e. the default settings
used in Pontarp et al. (2012).
}
\author{
Theo Pannetier
}
