% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_trait_gaps.R
\name{find_trait_gaps}
\alias{find_trait_gaps}
\title{Find gaps in trait values}
\usage{
find_trait_gaps(traits, trait_gap = default_trait_gap())
}
\arguments{
\item{traits}{a numeric vector, trait values \strong{in ascending order}.}

\item{trait_gap}{numeric, the width of a gap triggering speciation.}
}
\description{
Runs through an ordered vector of trait values, returns the positions of gaps
\verb{>= trait_gap} between consecutive values.
}
\author{
Théo Pannetier
}
