% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_carr_cap.R
\name{get_carr_cap}
\alias{get_carr_cap}
\title{Get the carrying capacity for a given trait value}
\usage{
get_carr_cap(
  trait_ind,
  trait_opt = default_trait_opt(),
  carr_cap_opt = default_carr_cap_opt(),
  carr_cap_width = default_carr_cap_width()
)
}
\arguments{
\item{trait_ind}{numeric. Trait value \eqn{z} of a focal individual.}

\item{trait_opt}{numeric. Optimal trait value \eqn{z_{opt}}, such that
\code{get_carr_cap(trait_opt) = carr_cap_opt}.}

\item{carr_cap_opt}{numeric. Maximum carrying capacity at \eqn{z = z_{opt}}.
Note that this corresponds to the maximum number of **competitors** for that
trait value, as the focal individual does not contribute to the carrying
capacity (see also \code{\link{get_fitness}}).}

\item{carr_cap_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{K}}. Controls
how fast the carrying capacity decays as the distance between a \code{trait}
and \code{trait_opt} increases.
\eqn{\sigma^{2}_{K}}}
}
\description{
Computes the carrying capacity experienced by an individual with trait value
\code{trait}.
}
\author{
Theo Pannetier
}
