% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitness.R
\name{get_fitness}
\alias{get_fitness}
\title{Get the fitness of an individual}
\usage{
get_fitness(
  traits_pop,
  growth_rate = default_growth_rate(),
  comp_width = default_comp_width(),
  trait_opt = default_trait_opt(),
  carr_cap_opt = default_carr_cap_opt(),
  carr_cap_width = default_carr_cap_width()
)
}
\arguments{
\item{traits_pop}{numeric vector of variable length, the trait values of
every individual in the population.}

\item{growth_rate}{numeric \eqn{>= 0}, the baseline growth rate in the
absence of competition.}

\item{comp_width}{numeric ≥ 0. \eqn{\sigma^{2}_{\alpha}}. Controls the
intensity of competition given a distance between two trait values.}

\item{trait_opt}{numeric. Optimal trait value \eqn{z_{opt}}, such that
\code{get_carr_cap(trait_opt) = carr_cap_opt}.}

\item{carr_cap_opt}{numeric. Maximum carrying capacity at \eqn{z = z_{opt}}.
Note that this corresponds to the maximum number of **competitors** for that
trait value, as the focal individual does not contribute to the carrying
capacity (see also \code{\link{get_fitness}}).}

\item{carr_cap_width}{numeric ≥ 0. \eqn{\sigma^{2}_{K}}. Controls how fast
the carrying capacity decays as the distance between a \code{trait} and
\code{trait_opt} increases.
\eqn{\sigma^{2}_{K}}}
}
\description{
Computes the fitness of an individual or population based on its trait value
and the trait values of other individuals in the population. The fitness here
corresponds to the average number of offspring an individual can get in a
generation, as sampled in a Poisson distribution.
}
\details{
The equation is a per-capita version of the Ricker model:
\deqn{G(z_i) = exp(r(1 - N_{eff} / K(z_i, z_opt)))}
where \eqn{N_{eff}} is the effective population size, i.e. sum of competitive
effects experienced by the focal individual for a given trait value:
\deqn{N_{eff} = sum_j(\alpha(z_i, z_j))}
}
\author{
Theo Pannetier
}
