% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_eff.R
\name{get_n_eff}
\alias{get_n_eff}
\title{Compute the effective population size}
\usage{
get_n_eff(traits_comm, comp_width = default_comp_width())
}
\arguments{
\item{traits_comm}{numeric vector of variable length, the trait values of
every individual in the community.}

\item{comp_width}{numeric \eqn{>= 0}. Width of the competition kernel.}
}
\description{
Computes \eqn{n_eff}, the effective population size experienced by an
individual.
}
\author{
Theo Pannetier
}
