% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_eff.R
\name{get_n_eff}
\alias{get_n_eff}
\title{Compute the effective population size}
\usage{
get_n_eff(traits_pop, comp_width = default_comp_width())
}
\arguments{
\item{traits_pop}{numeric vector of variable length, the trait values of
every individual in the population.}

\item{comp_width}{numeric ≥ 0. \eqn{\sigma^{2}_{\alpha}}. Controls the
intensity of competition given a distance between two trait values.}
}
\description{
Computes \eqn{n_eff}, the effective population size experienced by an
individual.
}
\author{
Theo Pannetier
}
