% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_n_eff}
\alias{get_n_eff}
\title{Compute the effective population size}
\arguments{
\item{z}{numeric vector, the trait values of all individuals in the
community.}

\item{competition_sd}{numeric \verb{>= 0}. Width of the competition kernel.}
}
\description{
Computes \code{n_eff}, the effective population size experienced by an
individual.
}
\details{
\code{n_eff} sums the competitive effects an individual receives from
every individual in the community, including the individual itself. It is
called effective population size because it is the size of the population
that is relevant for competition.
}
\author{
Thijs Janzen, Théo Pannetier
}
