% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generation_traits.R
\name{plot_generation_traits}
\alias{plot_generation_traits}
\title{Plot trait distribution in a generation}
\usage{
plot_generation_traits(sim_tbl, generation, binwidth = 0.02)
}
\arguments{
\item{sim_tbl}{a tibble containing the output of a comrad simulation, as
produced by \code{\link[=run_simulation]{run_simulation()}} and read by \code{\link[=read_sim_tbl]{read_sim_tbl()}}.}

\item{generation}{numeric, the index of the generation to plot}

\item{binwidth}{numeric, the width used for binning.}
}
\description{
Plots a histogram of the distribution of trait values in a single generation.
}
\author{
Théo Pannetier
}
