% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_population_size.R
\name{plot_population_size}
\alias{plot_population_size}
\title{Plot population size over generations}
\usage{
plot_population_size(sim_tbl)
}
\arguments{
\item{sim_tbl}{a tibble containing the output of a comrad simulation, as
produced by \code{\link[=run_simulation]{run_simulation()}} and read by \code{\link[=read_sim_tbl]{read_sim_tbl()}}.}
}
\description{
Plots the number of individuals in the population over time.
}
\author{
Théo Pannetier
}
