% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_population_trait_evolution.R
\name{plot_population_trait_evolution}
\alias{plot_population_trait_evolution}
\title{Plot the evolution of trait values in the population over generations}
\usage{
plot_population_trait_evolution(
  comrad_tbl,
  generation_range = c(0, Inf),
  xgrain = 10,
  ygrain = 0.01
)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}

\item{generation_range}{numeric vector with length 2, supplying the first and
last generation to plot from the dataset. Defaults to every generation in the
table}

\item{xgrain}{numeric, the width of  hexes on the x-axis (generations)}

\item{ygrain}{numeric, the width of  hexes on the y-axis (trait)}
}
\description{
Produces a hex-density plot of all trait values in the population over
generations.
}
\author{
Théo Pannetier
}
