% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(
  output_path,
  init_pop = default_init_pop(),
  nb_generations = 20,
  sampling_frequency = set_sampling_frequency(nb_generations),
  seed = default_seed(),
  growth_rate = default_growth_rate(),
  comp_width = default_comp_width(),
  trait_opt = default_trait_opt(),
  carr_cap_opt = default_carr_cap_opt(),
  carr_cap_width = default_carr_cap_width(),
  prob_mutation = default_prob_mutation(),
  mutation_sd = default_mutation_sd(),
  hpc_job_id = "local"
)
}
\arguments{
\item{output_path}{character, path to save the output file. If \code{NULL}, the
output is not saved and the population is returned at the end of the
simulation}

\item{init_pop}{a tibble containing the initial population.}

\item{nb_generations}{integer, the number of generations to run the
simulation for.}

\item{sampling_frequency}{numeric \code{> 0}, the frequency at which the
population is saved in the output.}

\item{seed}{numeric \code{> 0}, the integer seed to set for the random number
generator.}

\item{growth_rate}{numeric \eqn{>= 0}, the baseline growth rate in the
absence of competition.}

\item{comp_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{\alpha}}. Controls
the intensity of competition given a distance between two trait values.}

\item{trait_opt}{numeric. Optimal trait value \eqn{z_{opt}}, such that
\code{get_carr_cap(trait_opt) = carr_cap_opt}.}

\item{carr_cap_opt}{numeric. Maximum carrying capacity at \eqn{z = z_{opt}}.
Note that this corresponds to the maximum number of \strong{competitors} for that
trait value, as the focal individual does not contribute to the carrying
capacity (see also \code{\link{get_fitness}}).}

\item{carr_cap_width}{numeric \eqn{>= 0} 0. \eqn{\sigma^{2}_{K}}. Controls
how fast the carrying capacity decays as the distance between a \code{trait}
and \code{trait_opt} increases.
\eqn{\sigma^{2}_{K}}}

\item{prob_mutation}{numeric between 0 and 1, the probability that any new
individual is sampled with a mutation.}

\item{mutation_sd}{numeric \eqn{>= 0}, the standard deviation of the normal
distrbution in which the mutations are drawn.}

\item{hpc_job_id}{only relevant if run on a HPC, otherwise takes value
\code{"local"}. If supplied, the job ID will be included in the metadata.}
}
\description{
Run the competitive radiation simulation.
}
\details{
Output is registered in a .csv file with the following structure:
| t | z | runtime |
| --- | --- |--- |
| --- | --- |--- |
where each line is an individual (the entire population is recorded).
\code{t} denotes the generation counter, \code{z} is a trait value, and
\code{runtime} is the processing time elapsed during the current generation.
In addition, the output table is preceded by metadata. Skip the 14 first
lines to get the table proper.
}
\author{
Theo Pannetier
}
