% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(
  output_path,
  init_comm = default_init_comm(),
  nb_generations = 20,
  sampling_frequency = comrad::set_sampling_frequency(nb_generations),
  seed = default_seed(),
  growth_rate = default_growth_rate(),
  comp_width = default_comp_width(),
  trait_opt = default_trait_opt(),
  carr_cap_opt = default_carr_cap_opt(),
  carr_cap_width = default_carr_cap_width(),
  prob_mutation = default_prob_mutation(),
  mutation_sd = default_mutation_sd(),
  hpc_job_id = NULL
)
}
\arguments{
\item{output_path}{character, path to save the output file, which must be a
\code{.csv}. If \code{NULL}, the output is not saved and the final state of the
community is returned at the end of the simulation.}

\item{init_comm}{The initial community, must have the same tibble structure
as the \code{\link[=default_init_comm]{default_init_comm()}}, which contains 10 individuals with \code{z = 0}}

\item{nb_generations}{integer, the number of generations to run the
simulation for.}

\item{sampling_frequency}{numeric \code{> 0}, the frequency at which the
community is saved in the output. Default is \code{max(1, nb_generations / 100)}.}

\item{seed}{integer \code{> 0}, the seed to set for the random number
generator.}

\item{growth_rate}{numeric \verb{>= 0}, the baseline growth rate. Generations
being discrete, high values will cause chaos.}

\item{comp_width}{numeric \verb{>= 0}. Width of the competition kernel.}

\item{trait_opt}{numeric. The optimal trait value.
\code{get_carr_cap(trait_opt) = carr_cap_opt}.}

\item{carr_cap_opt}{numeric. Carrying capacity at the optimal trait value.}

\item{carr_cap_width}{numeric \verb{>= 0}.
Width of the carrying capacity kernel.}

\item{prob_mutation}{numeric between 0 and 1, the probability that any new
individual is sampled with a mutation.}

\item{mutation_sd}{numeric \verb{>= 0}, the standard deviation of the normal
distrbution from which mutations are drawn.}

\item{hpc_job_id}{used to record a job ID in the metadata, only relevant for
simulations run on a high-performance cluster. Otherwise takes value
\code{"local"}.}
}
\value{
Returns a table with a row corresponding to each individual, and five
columns: \code{t} is the generation time, \code{z} the individual's trait value,
\code{species} the name of the species it belongs to, and \code{ancestral_species} the
previous species it descends from.
If \code{output_path = NULL}, the community at the last generation is returned.
If the path to a \code{.csv} file is supplied, each sampled generation is appended
to the file. In the \code{.csv}, the table is preceded by 17 lines of metadata,
which are automatically ignored if the file is read with \code{\link[=read_comrad_tbl]{read_comrad_tbl()}}.
}
\description{
Run the competitive radiation simulation.
}
\author{
Théo Pannetier
}
