% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sampling_frequency.R
\name{set_sampling_frequency}
\alias{set_sampling_frequency}
\title{Set default sampling frequency}
\usage{
set_sampling_frequency(nb_generations)
}
\arguments{
\item{nb_generations}{integer, the number of generations to run during the
simulation.}
}
\description{
Returns an appropriate sampling frequency based on the order of magnitude of
\code{nb_generations}.
}
\details{
The sampling frequency is set as follows:
\itemize{
\item if \code{nb_generations < 1000}, sampling frequency is set to 1
(every generation).
\item if \code{nb_generations >= 1000}, sampling frequency is set to the
order of magnitude of \code{nb_generations} - 2, e.g. 2 for
\code{nb_generations >= 1000}, 3 for \code{nb_generations >= 10000}, etc.
}
}
\author{
Theo Pannetier
}
